move EnikshayCaseFactory to own file
